create user comment featuer and auth with Gate and delete user his  article and comments

+++ app/Http/Controllers/ArticleController.php
@@ -6,9 +6,15 @@ namespace App\Http\Controllers;
 use Illuminate\Http\Request;
 use App\Models\Article;
 use App\Models\Category;
+use Illuminate\Support\Facades\Gate;
 
 class ArticleController extends Controller
 {
+    public function __construct()
+    {
+        $this->middleware('auth')->except(["index", "detail"]);
+    }
+
     public function index()
     {
         $data = Article::latest()->paginate(5);
@@ -48,6 +54,7 @@ class ArticleController extends Controller
         $article->title = request()->title;
         $article->body = request()->body;
         $article->category_id = request()->category_id;
+        $article->user_id = auth()->id();
         $article->save();
 
         return redirect("articles");
@@ -85,8 +92,11 @@ class ArticleController extends Controller
     public function delete($id)
     {
         $article = Article::find($id);
-        $article->delete();
-
-        return redirect("/articles")->with("info", "Article Deleted");
+        if (Gate::allows("delete-article", $article)) {
+            $article->delete();
+            return redirect("/articles")->with("info", "Article Deleted");
+        } else {
+            return back()->with("info", "Unauthorize");
+        }
     }
 }
